+++ src/Router/RouteDispatcher.php
@@ -9,83 +9,85 @@
  * @author Arman Ag. <arman.ag@softberg.org>
  * @copyright Copyright (c) 2018 Softberg LLC (https://softberg.org)
  * @link http://quantum.softberg.org/
- * @since 2.9.5
+ * @since 2.9.7
  */
 
-namespace Quantum\Mvc;
+namespace Quantum\Router;
 
+use Quantum\Libraries\Encryption\Exceptions\CryptorException;
+use Quantum\Libraries\Database\Exceptions\DatabaseException;
+use Quantum\Libraries\Session\Exceptions\SessionException;
+use Quantum\Libraries\Config\Exceptions\ConfigException;
+use Quantum\Libraries\Csrf\Exceptions\CsrfException;
 use Quantum\Exceptions\ControllerException;
 use Quantum\Middleware\MiddlewareExecutor;
+use Quantum\Di\Exceptions\DiException;
 use Quantum\Handlers\ViewCacheHandler;
-use Quantum\Router\RouteController;
+use Quantum\Exceptions\BaseException;
 use Quantum\Libraries\Csrf\Csrf;
+use Quantum\Http\Request;
 use Quantum\Http\Response;
 use Quantum\Loader\Loader;
-use Quantum\Http\Request;
+use ReflectionException;
 use Quantum\Di\Di;
 
 class RouteDispatcher
 {
+
     /**
      * Handles the incoming HTTP request and generates a response.
-     *
-     * This includes:
-     * 1. Executing registered middleware.
-     * 2. Attempting to serve a cached view.
-     * 3. Handling routing through either a route callback or a controller action.
-     *
-     * Controller lifecycle hooks `__before` and `__after` will be invoked if defined.
-     * CSRF token verification is performed for applicable HTTP methods.
-     *
-     * @param Request  $request  The incoming HTTP request object.
-     * @param Response $response The HTTP response object to be sent.
-     *
+     * @param Request $request
+     * @param Response $response
      * @return void
+     * @throws ControllerException
+     * @throws DiException
+     * @throws BaseException
+     * @throws ConfigException
+     * @throws CsrfException
+     * @throws DatabaseException
+     * @throws CryptorException
+     * @throws SessionException
+     * @throws ReflectionException
      */
     public static function handle(Request $request, Response $response): void
     {
-        // 1. Apply middleware
-        [$request, $response] = (new MiddlewareExecutor())->execute($request, $response);
+        list($request, $response) = (new MiddlewareExecutor())->execute($request, $response);
 
-        // 2. Try serving from view cache
         $viewCacheHandler = new ViewCacheHandler();
         if ($viewCacheHandler->serveCachedView(route_uri(), $response)) {
             return;
         }
 
-        // 3. Route callback or controller handling
         $callback = route_callback();
 
         if ($callback) {
             call_user_func_array($callback, self::getArgs($callback));
-        } else {
-            $controller = self::getController();
-            $action = self::getAction($controller);
+            return;
+        }
 
-            if ($controller->csrfVerification && in_array($request->getMethod(), Csrf::METHODS)) {
-                csrf()->checkToken($request);
-            }
+        $controller = self::getController();
+        $action = self::getAction($controller);
 
-            if (method_exists($controller, '__before')) {
-                call_user_func_array([$controller, '__before'], self::getArgs([$controller, '__before']));
-            }
+        if ($controller->csrfVerification && in_array($request->getMethod(), Csrf::METHODS)) {
+            csrf()->checkToken($request);
+        }
 
-            call_user_func_array([$controller, $action], self::getArgs([$controller, $action]));
+        if (method_exists($controller, '__before')) {
+            call_user_func_array([$controller, '__before'], self::getArgs([$controller, '__before']));
+        }
 
-            if (method_exists($controller, '__after')) {
-                call_user_func_array([$controller, '__after'], self::getArgs([$controller, '__after']));
-            }
+        call_user_func_array([$controller, $action], self::getArgs([$controller, $action]));
+
+        if (method_exists($controller, '__after')) {
+            call_user_func_array([$controller, '__after'], self::getArgs([$controller, '__after']));
         }
     }
 
     /**
      * Loads and returns the current route's controller instance.
-     *
-     * Will throw a ControllerException if the controller is not found or not defined properly.
-     *
-     * @return RouteController The loaded controller instance.
-     *
-     * @throws ControllerException
+     * @return RouteController
+     * @throws DiException
+     * @throws ReflectionException
      */
     private static function getController(): RouteController
     {
@@ -105,20 +107,16 @@ class RouteDispatcher
     }
 
     /**
-     * Retrieves the current route's action (method name) for the controller.
-     *
-     * @param RouteController $controller The controller instance to check against.
-     *
-     * @return string|null The action method name, or null if none is defined.
-     *
-     * @throws ControllerException If the action method does not exist in the controller.
+     * Retrieves the current route's action for the controller.
+     * @param RouteController $controller
+     * @return string|null
+     * @throws ControllerException
      */
-
     private static function getAction(RouteController $controller): ?string
     {
         $action = current_action();
 
-        if ($action && !method_exists($controller, $action)) {
+        if (!$action || !method_exists($controller, $action)) {
             throw ControllerException::actionNotDefined($action);
         }
 
@@ -127,10 +125,10 @@ class RouteDispatcher
 
     /**
      * Resolves and returns the arguments for a given callable using dependency injection.
-     *
-     * @param callable $callable The function or method to be invoked.
-     *
-     * @return array The resolved arguments for the callable.
+     * @param callable $callable
+     * @return array
+     * @throws DiException
+     * @throws ReflectionException
      */
     private static function getArgs(callable $callable): array
     {
@@ -139,8 +137,7 @@ class RouteDispatcher
 
     /**
      * Retrieves the route parameters from the current route.
-     *
-     * @return array An array of parameter values for the current route.
+     * @return array
      */
     private static function routeParams(): array
     {
